Consolidate container helper functions
