+++ ear-production-suite-plugins/lib/CMakeLists.txt
@@ -131,6 +131,7 @@ set(EAR_BASE_HEADERS
 	include/scene_store.hpp
 	include/communication/metadata_thread.hpp
 	include/auto_mode_controller.hpp
+	${EPS_SHARED_DIR}/helper/container_helpers.hpp
 	${EPS_SHARED_DIR}/speaker_setups.hpp
 	)
