+++ ear-production-suite-plugins/lib/include/binaural_monitoring_backend.hpp
@@ -16,36 +16,6 @@
 
 using ConnId = std::string;
 
-template <typename Key, typename Value>
-std::optional<Value> getFromMap(std::map<Key, Value>& targetMap, Key key) {
-  auto it = targetMap.find(key);
-  if (it == targetMap.end()) return std::optional<Value>();
-  return std::optional<Value>(it->second);
-}
-
-template <typename Key, typename Value>
-void removeFromMap(std::map<Key, Value>& targetMap, Key key) {
-  auto it = targetMap.find(key);
-  if (it != targetMap.end()) targetMap.erase(it);
-}
-
-template <typename Key, typename Value>
-Value* getValuePointerFromMap(std::map<Key, Value>& targetMap, Key key) {
-  auto it = targetMap.find(key);
-  if (it == targetMap.end()) return nullptr;
-  return &(it->second);
-}
-
-template <typename Key, typename Value>
-void setInMap(std::map<Key, Value>& targetMap, Key key, Value value) {
-  auto it = targetMap.find(key);
-  if (it == targetMap.end()) {
-    targetMap.insert(std::make_pair(key, value));
-  } else {
-    it->second = value;
-  }
-}
-
 namespace ear {
 namespace plugin {
 namespace ui {
