+++ ear-production-suite-plugins/plugins/binaural_monitoring/CMakeLists.txt
@@ -37,6 +37,7 @@ set(HEADERS_BINAURAL_MONITORING
 	
 	${EPS_SHARED_DIR}/components/look_and_feel/slider.hpp
 	
+	${EPS_SHARED_DIR}/helper/container_helpers.hpp
 	${EPS_SHARED_DIR}/helper/graphics.hpp
 	${EPS_SHARED_DIR}/helper/multi_async_updater.h
 	${EPS_SHARED_DIR}/helper/properties_file.hpp
