+++ reaper-adm-extension/src/reaper_adm/pcmwriterfactory.cpp
@@ -7,9 +7,6 @@
 
 using namespace admplug;
 
-namespace {
-}
-
 RoutingWriterFactory::RoutingWriterFactory()
 {
